Created config file for log_daemon settings. Created unit tests for log_daemon.

+++ bin/log_daemon.py
@@ -1,42 +1,53 @@
 import daemon
 import os
 import time
-import configparser
 import logging
-from zipfile import ZipFile
+import configparser
 from impl.dual_logging import get_last_log_ID
 from utils.condor_utils import *
 
 
 class log_manager:
     
-    def __init__(self):
-        ## CONFIGURATION VARIABLES: (to be written in a configuration file and assigned here)
+    def __init__(self,cfg):
         #  dT time after which a job is considered historic (Folder is zipped and moved to "old")
-        self.dT = 43200.0 # 1 days
+        self.dT = float(cfg['Time variables']['Historic after'])
         # dt time after which we considered a job UNSUCCESSFULL after being removed from the condor_q (job_control_daemon gets killed)
-        self.dt = 3600.0 # 1 h
+        self.dt = float(cfg['Time variables']['Unsuccessfull after'])
         # dH time that a job can be on hold before being UNSUCCESSFULL
-        self.dH = 180000.0 # 5 days
+        self.dH = float(cfg['Time variables']['On hold untill'])
         # how often we run the daemon
-        self.running_freq = 60 #s
+        self.running_freq = float(cfg['Time variables']['Running_freq'])
         # global log file
-        self.logfile ="/opt/IDEAL-1.1test/data/logs/IDEAL_general_logs.log"
+        self.logfile = cfg['Paths']['Global logfile']
         # global control file for cleaning up and debug purposes
-        self.cfg_log_file = "/opt/IDEAL-1.1test/data/logs/ideal_history.cfg"
+        self.cfg_log_file = cfg['Paths']['Cfg_log_file']
+        # log daemon logs
+        self.log_daemon_logs = cfg['Paths']['Log_daemon_logs']
         # Directory for historical jobs (completed and failed)
-        self.completed_dir = "/opt/IDEAL-1.1test/data/old/completed"
-        self.failed_dir = "/opt/IDEAL-1.1test/data/old/failed"
+        self.completed_dir = cfg['Paths']['Completed_dir']
+        self.failed_dir = cfg['Paths']['Failed_dir']
         # Archive for log files
-        self.logs_folder = "/opt/IDEAL-1.1test/data/logs"
+        self.logs_folder = cfg['Paths']['Logs_folder']
         # dictionary of possible condor job status
-        self.job_status_dict = {'submission_err':'SUBMISSION ERROR','unsuccessfull':'UNSUCCESSFULL','done':'DONE',
-                           'killed_by_log_daem':'KILLED BY LOG DAEMON','checking':'BEING CHECKED'}
+        self.job_status_dict = cfg['Job status']
         # List possible epilogue of condor jobs
-        self.end_status = ['SUBMISSION ERROR','UNSUCCESSFULL', 'DONE', 'KILLED BY LOG DAEMON']
+        self.end_status = [self.job_status_dict['Submission_err'],self.job_status_dict['Unsuccessfull'], self.job_status_dict['Done'], self.job_status_dict['Killed_by_log_daem']]
         
         self.parser = None
+        self.log = self.get_log_file(self.log_daemon_logs,'%(asctime)s - %(levelname)s - %(message)s')
         
+        
+    def get_log_file(self,log_daemon_logs,formatt):
+        formatter = logging.Formatter(formatt)
+        handler = logging.FileHandler(log_daemon_logs)        
+        handler.setFormatter(formatter)
+    
+        logger = logging.getLogger()
+        logger.setLevel(logging.DEBUG)
+        logger.addHandler(handler)
+    
+        return logger
         
     # read files is called outside the class to allow easier unittesting    
     def read_files(self): 
@@ -55,7 +66,7 @@ class log_manager:
     	last_ID_log = int(get_last_log_ID())
 
         # Get daemons. Read daemons before updating config!
-    	log.info("Get job daemons")
+    	self.log.info("Get job daemons")
     	self.daemons = get_job_daemons("job_control_daemon.py")
             
     	# Create new sections for the newly added IDs
@@ -64,14 +75,14 @@ class log_manager:
     		self.add_id_sections(id_range)
     		# Check for free running daemons
             # NOTE: done here to avoid checking on not up to date cfg file
-    		log.info("Find and kill daemons for failed or untracked jobs")
+    		self.log.info("Find and kill daemons for failed or untracked jobs")
     		self.kill_untracked_daemons(self.daemons)
             
     
     def update_log_file(self):
         	# Wake up to work 
         # Get job status
-        	log.info("Reading condor queue")
+        	self.log.info("Reading condor queue")
         	self.all_jobs = get_jobs_status()
             
         	parser = self.parser
@@ -80,16 +91,16 @@ class log_manager:
         			job_age = get_job_age(parser[i]['Submission date'],"%Y-%m-%d %H:%M:%S")
         			# Update the status for recent sections (date - submission date < dT)
         			if job_age < self.dT:
-        				log.info("Checking section {}".format(i))
+        				self.log.info("Checking section {}".format(i))
         				# ideal status
-        				log.info("Update ideal status")
+        				self.log.info("Update ideal status")
         				self.update_ideal_status(parser[i])
         				# condor status
         				if parser[i]['Condor status'] not in self.end_status:
-        					log.info("Update condor status")
+        					self.log.info("Update condor status")
         					self.update_job_status(parser[i],self.all_jobs)
         				# job control daemon
-        				log.info("Update daemon status")
+        				self.log.info("Update daemon status")
         				self.update_job_daemon_status(parser[i],self.daemons)
         				# kill daemons for unsuccessful jobs
         				self.kill_running_daemons(parser[i])
@@ -98,11 +109,11 @@ class log_manager:
         				self.cleanup_workdir(parser[i],self.completed_dir,self.failed_dir)
         		else:
         			parser[i]['Condor status'] = self.job_status_dict['submission_err']
-        			log.info("Submission error. Going to archive the job as failed")
+        			self.log.info("Submission error. Going to archive the job as failed")
         			self.cleanup_workdir(parser[i],self.completed_dir,self.failed_dir)
             
         # Clean up historical logs
-        	log.info("Zip historical log files")
+        	self.log.info("Zip historical log files")
         	self.archive_logs()
         	
         	# Dump the new configuration file
@@ -110,60 +121,60 @@ class log_manager:
         			self.parser.write(configfile)
                     
         # Sleep
-        	log.info("Going to sleep for {} s\n\n".format(self.running_freq))
+        	self.log.info("Going to sleep for {} s\n\n".format(self.running_freq))
         	time.sleep(self.running_freq)
-        	log.info("Waking up to work")
+        	self.log.info("Waking up to work")
         
     def update_ideal_status(self,pars_sec):
         cfg = configparser.ConfigParser()
         cfg.read(pars_sec['Simulation settings'])
         pars_sec['Status'] = cfg['DEFAULT']['status']
-        log.debug("status: {}".format(pars_sec['Status']))
+        self.log.debug("status: {}".format(pars_sec['Status']))
     
 
     def update_job_status(self,pars_sec, all_jobs):
         if pars_sec['Condor id'] in all_jobs:  # job is in condor_q (running, idle, done, hold)
             job_id = pars_sec['Condor id']
             pars_sec['Condor status'] = str(all_jobs[job_id])
-            log.info("Job in condor queue")
+            self.log.info("Job in condor queue")
             if job_on_hold(all_jobs,job_id):
                 if 'On hold since' not in pars_sec:
-                    log.debug("Job was put on hold")
+                    self.log.debug("Job was put on hold")
                     pars_sec['On hold since'] = time.strftime("%Y-%m-%d %H:%M:%S")
                 if condor_check_run() == 0: # no problems with condor
                     # try to release
-                    log.debug("Try to release job")
+                    self.log.debug("Try to release job")
                     ret = release_condor_job(job_id)
                     if ret == 0:
                         pars_sec['On hold since'] = 'Released successfully'
-                        log.debug("Job released with exit code 0")
+                        self.log.debug("Job released with exit code 0")
                     else: # if it was too long on hold, remove and mark as UNSUCCESSFULL
                         if get_job_age(pars_sec['On hold since'],"%Y-%m-%d %H:%M:%S") > self.dH:
-                            log.debug("Tried to release for {} sand failed. Job will be removed from queue".format(self.dH))
+                            self.log.debug("Tried to release for {} sand failed. Job will be removed from queue".format(self.dH))
                             remove_condor_job(job_id) 
                             pars_sec['Condor status'] = self.job_status_dict['killed_by_log_daem']
         else:            
             if pars_sec['Status'] == 'FINISHED':
-                log.info("Job not in the queue because of successful termination")
+                self.log.info("Job not in the queue because of successful termination")
                 pars_sec['Condor status'] = self.job_status_dict['done']
             else:
                  if pars_sec['Condor status'] != self.job_status_dict['checking']:
                      pars_sec['Condor status'] = self.job_status_dict['checking']
                      pars_sec['Last checked'] = time.strftime("%Y-%m-%d %H:%M:%S")
-                     log.info("Job not in the queue but not terminated. Wait.")
+                     self.log.info("Job not in the queue but not terminated. Wait.")
                  else:
                      if get_job_age(pars_sec['Last checked'],"%Y-%m-%d %H:%M:%S") > self.dt: # if the job has been checked for longer then an hour
-                         log.info("Job not in the queue and still not terminated after {} s. Marked as failed.".format(self.dt))
+                         self.log.info("Job not in the queue and still not terminated after {} s. Marked as failed.".format(self.dt))
                          pars_sec['Condor status'] = self.job_status_dict['unsuccessfull']
-        log.debug("Condor status: {}".format(pars_sec['Condor status']))
+        self.log.debug("Condor status: {}".format(pars_sec['Condor status']))
                          
     def update_job_daemon_status(self,pars_sec,daemons):
         if pars_sec['Work_dir'] in daemons:
-            log.info("Daemon is running")
+            self.log.info("Daemon is running")
             pid = daemons[pars_sec['Work_dir']]
             pars_sec['Job control daemon'] = 'Running with pid {}'.format(pid)
         elif pars_sec['Job control daemon'] != 'Daemon killed': 
-            log.info("Daemon not running. Assumed successful termination")
+            self.log.info("Daemon not running. Assumed successful termination")
             pars_sec['Job control daemon'] = 'Daemon successfully finished'
     
     def is_daemon_to_kill(self,pars_sec):
@@ -173,12 +184,16 @@ class log_manager:
         else:
             return False
     
-    def kill_running_daemons(self,pars_sec):
+    def kill_running_daemons(self,pars_sec,test=False):
         if self.is_daemon_to_kill(pars_sec):
             pid = pars_sec['Job control daemon'].split(" ")[-1]
-            ret = kill_process(pid)
+            if not test:
+                ret = kill_process(pid)
             pars_sec['Job control daemon'] = 'Daemon killed'
-            log.info("Daemon for job {} killed by program".format(self.pars_sec['Condor id']))
+            self.log.info("Daemon for job {} killed by program".format(pars_sec['Condor id']))
+            
+            if test: return pid
+            
             
     def kill_untracked_daemons(self,daemons,test=False):
         # kill daemons not associated to any job
@@ -189,7 +204,7 @@ class log_manager:
                 pid = d[1]
                 if not test:
                     ret = kill_process(pid)
-                    log.info("Daemon killed because not connected to any tracked job.\nWorkdir: {}".format(d[0]))
+                    self.log.info("Daemon killed because not connected to any tracked job.\nWorkdir: {}".format(d[0]))
                 else: a.append(pid)
                     
         if test: return a
@@ -218,34 +233,34 @@ class log_manager:
         if not os.path.isdir(completed_dir):
             os.makedirs(completed_dir)
             
-        log.info("Job considered historical. Going to zip output and work dir")
+        self.log.info("Job considered historical. Going to zip output and work dir")
         
         workdir = pars_sec['Work_dir']
         data = workdir.split("work")[0] 
         workdir = workdir.split("/rungate")[0]
-        log.debug("Working dir: {}".format(workdir))
+        self.log.debug("Working dir: {}".format(workdir))
         # get output dir
         fname =  workdir.split("work")[1]
         output = data+"output"+fname
-        log.debug("Output dir: {}".format(output))
+        self.log.debug("Output dir: {}".format(output))
         
         base_out = data + "--output" # destination file for output dir
         base_work = data + "--work" # destination file for work dir
         
-        log.info("Zipping {0} in {1}. Removing {0}".format(output,base_out+".zip"))
+        self.log.info("Zipping {0} in {1}. Removing {0}".format(output,base_out+".zip"))
         zip_and_clean_folder(base_out,output)
-        log.info("Zipping {0} in {1}. Removing {0}".format(workdir,base_work+".zip"))
+        self.log.info("Zipping {0} in {1}. Removing {0}".format(workdir,base_work+".zip"))
         zip_and_clean_folder(base_work,workdir)
         
         # make zip file for the two previouslzy zipped files   
         if pars_sec['Status'] == 'FINISHED':
             os.chdir(data)
             zip_files(completed_dir+fname+'.zip',['--output.zip','--work.zip'])
-            log.info("Archiving BOTH in {}".format(completed_dir+fname+'.zip'))
+            self.log.info("Archiving BOTH in {}".format(completed_dir+fname+'.zip'))
         else:
             os.chdir(data)
             zip_files(failed_dir+fname+'.zip',['--output.zip','--work.zip'])
-            log.info("Archiving in {}".format(failed_dir+fname+'.zip'))
+            self.log.info("Archiving in {}".format(failed_dir+fname+'.zip'))
             
         os.remove(base_out+'.zip')    
         os.remove(base_work+'.zip')
@@ -290,7 +305,7 @@ class log_manager:
      
         for i in range(len(id_range)):
             self.config_template(self.parser,id_range[i],work_dirs[i],dates[i],condor_ids[i],settings[i])
-            log.info("Added section with ID: {}".format(id_range[i]))
+            self.log.info("Added section with ID: {}".format(id_range[i]))
         
     		
     def config_template(self,config,idealID,workdir,date,condor_id,settings):
@@ -307,22 +322,14 @@ class log_manager:
 
 if __name__ == '__main__':
     
-    # Create log manager class:
-    manager = log_manager()
+    cfg_parser = configparser.ConfigParser()
+    cfg_parser.read("/opt/IDEAL-1.1test/cfg/log_daemon.cfg")
     
     with daemon.DaemonContext():
-        # Create logging system after demonizing as it does not like to be daemonized
-        # Log daemon log file
-    	global log
-    	log_daemon_logs = "/opt/IDEAL-1.1test/data/logs/log_daemon.log"
-    	formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
-    	handler = logging.FileHandler(log_daemon_logs)        
-    	handler.setFormatter(formatter)
-    	log = logging.getLogger()
-    	log.setLevel(logging.DEBUG)
-    	log.addHandler(handler)
+        # Create log manager class:
+        manager = log_manager(cfg_parser)
         
-    	while True:  # To stop run bin/stop_log_daemon
+        while True:  # To stop run bin/stop_log_daemon
             # Read main log file and update config file with new entries
             manager.read_files()
             manager.update_log_file()
